chore: swap app directories
