--- src/app-basic/add/page.tsx
@@ -1,7 +1,0 @@
-export default function AddPlantPage() {
-  return (
-    <div className="p-4">
-      <h1 className="text-xl font-bold">Add a Plant</h1>
-    </div>
-  );
-}
