--- src/app-basic/layout.tsx
@@ -1,13 +1,0 @@
-import type { ReactNode } from "react";
-import { Providers } from "./providers";
-import "./globals.css";
-
-export default function RootLayout({ children }: { children: ReactNode }) {
-  return (
-    <html lang="en">
-      <body>
-        <Providers>{children}</Providers>
-      </body>
-    </html>
-  );
-}
