--- src/app-basic/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return <div className="p-4">Loading...</div>;
-}
