--- src/app-basic/page.tsx
@@ -1,4 +1,0 @@
-import TodayPage from "./today/page";
-export default function Page() {
-  return <TodayPage />;
-}
