--- src/app-basic/plants/page.tsx
@@ -1,7 +1,0 @@
-export default function PlantsPage() {
-  return (
-    <div className="p-4">
-      <h1 className="text-xl font-bold">Plants</h1>
-    </div>
-  );
-}
