--- src/app-basic/preview/page.tsx
@@ -1,7 +1,0 @@
-export default function PreviewPage() {
-  return (
-    <div className="p-4">
-      <h1 className="text-xl font-bold">Preview</h1>
-    </div>
-  );
-}
