--- src/app-basic/providers.tsx
@@ -1,6 +1,0 @@
-"use client";
-import type { ReactNode } from "react";
-
-export function Providers({ children }: { children: ReactNode }) {
-  return <>{children}</>;
-}
