--- src/app-basic/stats/page.tsx
@@ -1,7 +1,0 @@
-export default function StatsPage() {
-  return (
-    <div className="p-4">
-      <h1 className="text-xl font-bold">Stats</h1>
-    </div>
-  );
-}
