--- src/app-basic/today/page.tsx
@@ -1,8 +1,0 @@
-export default function TodayPage() {
-  return (
-    <div className="p-4">
-      <h1 className="text-xl font-bold">Today</h1>
-      <p className="mt-2">No tasks yet.</p>
-    </div>
-  );
-}
