+++ src/app-old/api/ai-care/route.ts
