+++ src/app-old/api/plants/[id]/route.ts
