+++ src/app-old/api/plants/route.ts
