+++ src/app-old/api/stats/route.ts
