+++ src/app-old/loading.tsx
@@ -1,0 +1,8 @@
+// src/app/loading.tsx
+export default function Loading() {
+  return (
+    <div className="min-h-dvh grid place-items-center p-8">
+      <div className="text-sm text-muted-foreground">Loading…</div>
+    </div>
+  );
+}
