+++ src/app-old/page.tsx
@@ -1,0 +1,8 @@
+import TodayPage, { revalidate } from "./today/page";
+
+export { revalidate };
+
+export default function Page() {
+  return <TodayPage />;
+}
+
