+++ src/app-old/plants/[id]/detail-view.tsx
