+++ src/app/page.tsx
@@ -1,8 +1,4 @@
-import TodayPage, { revalidate } from "./today/page";
-
-export { revalidate };
-
+import TodayPage from "./today/page";
 export default function Page() {
   return <TodayPage />;
 }
-
