+++ src/app/providers.tsx
@@ -1,13 +1,6 @@
-'use client';
-
-import { ThemeProvider } from "next-themes";
+"use client";
 import type { ReactNode } from "react";
 
 export function Providers({ children }: { children: ReactNode }) {
-  return (
-    <ThemeProvider attribute="class" defaultTheme="light" enableSystem={false}>
-      {children}
-    </ThemeProvider>
-  );
+  return <>{children}</>;
 }
-
